Make host_metrics generator a default configuration of the more flexible metrics generator
