+++ receiver/telemetrygeneratorreceiver/generators.go
@@ -28,6 +28,9 @@ const (
 	// generatorTypeLogs is the generator type for logs
 	generatorTypeLogs generatorType = "logs"
 
+	// generatorTypeMetrics is the generator type for generic metrics
+	generatorTypeMetrics generatorType = "metrics"
+
 	// generatorTypeHostMetrics is the generator type for host metrics
 	generatorTypeHostMetrics generatorType = "host_metrics"
 
@@ -75,6 +78,8 @@ func newMetricsGenerators(cfg *Config, logger *zap.Logger) []metricGenerator {
 	var generators []metricGenerator
 	for _, gen := range cfg.Generators {
 		switch gen.Type {
+		case generatorTypeMetrics:
+			generators = append(generators, newMetricsGenerator(gen, logger))
 		case generatorTypeHostMetrics:
 			generators = append(generators, newHostMetricsGenerator(gen, logger))
 		case generatorTypeOTLP:
