+++ receiver/telemetrygeneratorreceiver/host_metrics_generator.go
@@ -15,116 +15,56 @@
 package telemetrygeneratorreceiver //import "github.com/observiq/bindplane-agent/receiver/telemetrygeneratorreceiver"
 
 import (
-	"math"
-	"math/rand"
-
-	"go.opentelemetry.io/collector/pdata/pcommon"
-	"go.opentelemetry.io/collector/pdata/plog"
-	"go.opentelemetry.io/collector/pdata/pmetric"
-	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.uber.org/zap"
 )
 
 // hostMetricsGenerator is a generator for host metrics. It generates a sampling of host metrics
 // emulating the Host Metrics receiver: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/hostmetricsreceiver
-type hostMetricsGenerator struct {
-	otlpGenerator
-	// dataPointUpdaters is a list of functions that update the data points with new values
-	// for each batch of metrics that are generated
-	dataPointUpdaters []func()
+
+var defaultConfig = GeneratorConfig{
+	Type: "host_metrics",
+	AdditionalConfig: map[string]any{
+		"metrics": []any{
+			map[string]any{"name": "system.memory.utilization", "value_min": 0, "value_max": 100, "type": "Gauge", "unit": "1", "attributes": map[string]any{"state": "slab_unreclaimable"}},
+			map[string]any{"name": "system.memory.utilization", "value_min": 0, "value_max": 100, "type": "Gauge", "unit": "1", "attributes": map[string]any{"state": "cached"}},
+			map[string]any{"name": "system.memory.utilization", "value_min": 0, "value_max": 100, "type": "Gauge", "unit": "1", "attributes": map[string]any{"state": "slab_reclaimable"}},
+			map[string]any{"name": "system.memory.utilization", "value_min": 0, "value_max": 100, "type": "Gauge", "unit": "1", "attributes": map[string]any{"state": "buffered"}},
+			map[string]any{"name": "system.memory.usage", "value_min": 100000, "value_max": 1000000000, "type": "Sum", "unit": "By", "attributes": map[string]any{"state": "buffered"}},
+			map[string]any{"name": "system.memory.usage", "value_min": 100000, "value_max": 1000000000, "type": "Sum", "unit": "By", "attributes": map[string]any{"state": "slab_reclaimable"}},
+			map[string]any{"name": "system.memory.usage", "value_min": 100000, "value_max": 1000000000, "type": "Sum", "unit": "By", "attributes": map[string]any{"state": "slab_unreclaimable"}},
+			map[string]any{"name": "system.memory.usage", "value_min": 100000, "value_max": 1000000000, "type": "Sum", "unit": "By", "attributes": map[string]any{"state": "cached"}},
+			map[string]any{"name": "system.cpu.load_average.1m", "value_min": 0, "value_max": 1, "type": "Gauge", "unit": "{thread}"},
+			map[string]any{"name": "system.filesystem.usage", "value_min": 0, "value_max": 15616700416, "type": "Sum", "unit": "By", "attributes": map[string]any{"device": "/dev/vda1", "mode": "rw", "mountpoint": "/etc/hosts", "state": "reserved", "type": "ext4"}},
+			map[string]any{"name": "system.filesystem.usage", "value_min": 0, "value_max": 15616700416, "type": "Sum", "unit": "By", "attributes": map[string]any{"device": "/dev/vda1", "mode": "rw", "mountpoint": "/etc/hosts", "state": "free", "type": "ext4"}},
+			map[string]any{"name": "system.filesystem.utilization", "value_min": 0, "value_max": 1, "type": "Gauge", "unit": "1", "attributes": map[string]any{"device": "/dev/vda1", "mode": "rw", "mountpoint": "/etc/hosts", "state": "free", "type": "ext4"}},
+			map[string]any{"name": "system.filesystem.utilization", "value_min": 0, "value_max": 1, "type": "Gauge", "unit": "1", "attributes": map[string]any{"device": "/dev/vda1", "mode": "rw", "mountpoint": "/etc/hosts", "state": "free", "type": "ext4"}},
+			map[string]any{"name": "system.network.packets", "value_min": 0, "value_max": 1000000, "type": "Sum", "unit": "{packets}", "attributes": map[string]any{"device": "eth0", "direction": "receive"}},
+			map[string]any{"name": "system.network.packets", "value_min": 0, "value_max": 1000000, "type": "Sum", "unit": "{packets}", "attributes": map[string]any{"device": "eth0", "direction": "send"}},
+			map[string]any{"name": "system.network.io", "value_min": 0, "value_max": 100000000, "type": "Sum", "unit": "By", "attributes": map[string]any{"device": "eth0", "direction": "send"}},
+			map[string]any{"name": "system.network.io", "value_min": 0, "value_max": 100000000, "type": "Sum", "unit": "By", "attributes": map[string]any{"device": "eth0", "direction": "receive"}},
+			map[string]any{"name": "system.network.errors", "value_min": 0, "value_max": 1000, "type": "Sum", "unit": "{errors}", "attributes": map[string]any{"device": "eth0", "direction": "receive"}},
+			map[string]any{"name": "system.network.errors", "value_min": 0, "value_max": 1000, "type": "Sum", "unit": "{errors}", "attributes": map[string]any{"device": "eth0", "direction": "transmit"}},
+			map[string]any{"name": "system.network.dropped", "value_min": 0, "value_max": 1000, "type": "Sum", "unit": "{packets}", "attributes": map[string]any{"device": "eth0", "direction": "transmit"}},
+			map[string]any{"name": "system.network.dropped", "value_min": 0, "value_max": 1000, "type": "Sum", "unit": "{packets}", "attributes": map[string]any{"device": "eth0", "direction": "receive"}},
+			map[string]any{"name": "system.network.conntrack.max", "value_min": 65536, "value_max": 65536, "type": "Sum", "unit": "{entries}"},
+			map[string]any{"name": "system.network.conntrack.count", "value_min": 8, "value_max": 64, "type": "Sum", "unit": "{entries}"},
+			map[string]any{"name": "system.network.connections", "value_min": 0, "value_max": 64, "type": "Sum", "unit": "{connections}", "attributes": map[string]any{"protocol": "tcp", "state": "ESTABLISHED"}},
+			map[string]any{"name": "system.network.connections", "value_min": 0, "value_max": 64, "type": "Sum", "unit": "{connections}", "attributes": map[string]any{"protocol": "tcp", "state": "LISTEN"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "user", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "system", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "idle", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "interrupt", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "nice", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "softirq", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "steal", "cpu": "cpu0"}},
+			map[string]any{"name": "system.cpu.time", "value_min": 0, "value_max": 10000, "type": "Sum", "unit": "s", "attributes": map[string]any{"state": "wait", "cpu": "cpu0"}},
+		},
+	},
 }
 
 func newHostMetricsGenerator(cfg GeneratorConfig, logger *zap.Logger) metricGenerator {
-
-	g := &hostMetricsGenerator{
-		otlpGenerator: otlpGenerator{
-			cfg:     cfg,
-			logger:  logger,
-			logs:    plog.NewLogs(),
-			metrics: pmetric.NewMetrics(),
-			traces:  ptrace.NewTraces(),
-		},
-	}
-	// Load up Resources attributes
-
-	newResource := g.metrics.ResourceMetrics().AppendEmpty()
-	err := newResource.Resource().Attributes().FromRaw(cfg.ResourceAttributes)
-	if err != nil {
-		// should be caught in validation
-		logger.Error("Error setting resource attributes in host_metrics", zap.Error(err))
-	}
-
-	metrics := cfg.AdditionalConfig["metrics"].([]any)
-	newScope := newResource.ScopeMetrics().AppendEmpty()
-	for _, m := range metrics {
-		metric := m.(map[string]any)
-		var attributes map[string]any
-		// attributes are optional
-		if attr, ok := metric["attributes"]; ok {
-			attributes = attr.(map[string]any)
-		}
-
-		metricType := metric["type"].(string)
-
-		name := metric["name"].(string)
-		valueMin := metric["value_min"].(int)
-		valueMax := metric["value_max"].(int)
-		unit := metric["unit"].(string)
-
-		newMetric := newScope.Metrics().AppendEmpty()
-		newMetric.SetUnit(unit)
-		newMetric.SetName(name)
-
-		// all the host metrics are either sums or gauges
-		switch metricType {
-		case "Gauge":
-			newMetric.SetEmptyGauge()
-			dp := newMetric.Gauge().DataPoints().AppendEmpty()
-			dp.SetTimestamp(pcommon.NewTimestampFromTime(getCurrentTime()))
-			dp.SetStartTimestamp(pcommon.NewTimestampFromTime(getCurrentTime()))
-
-			err = dp.Attributes().FromRaw(attributes)
-			if err != nil {
-				// should be caught in validation
-				logger.Error("Error setting attributes in host_metrics", zap.String("name", name), zap.Error(err))
-			}
-
-			// All the host metric Gauges are float64
-			g.dataPointUpdaters = append(g.dataPointUpdaters, func() { dp.SetDoubleValue(getRandomFloat64(valueMin, valueMax)) })
-		case "Sum":
-			newMetric.SetEmptySum()
-			dp := newMetric.Sum().DataPoints().AppendEmpty()
-			dp.SetTimestamp(pcommon.NewTimestampFromTime(getCurrentTime()))
-			dp.SetStartTimestamp(pcommon.NewTimestampFromTime(getCurrentTime()))
-			err = dp.Attributes().FromRaw(attributes)
-			if err != nil {
-				// should be caught in validation
-				logger.Error("Error setting attributes in host_metrics", zap.String("name", name), zap.Error(err))
-			}
-			switch unit {
-			case "s":
-				g.dataPointUpdaters = append(g.dataPointUpdaters, func() { dp.SetDoubleValue(float64(math.Trunc(getRandomFloat64(valueMin, valueMax)*100)) / 100) })
-			default:
-				g.dataPointUpdaters = append(g.dataPointUpdaters, func() { dp.SetIntValue(int64(getRandomFloat64(valueMin, valueMax))) })
-			}
-		}
-	}
-
-	g.adjustMetricTimes()
-
+	// ignore the config, which only contains the type and resource attributes
+	defaultConfig.ResourceAttributes = cfg.ResourceAttributes
+	g := newMetricsGenerator(defaultConfig, logger)
 	return g
-}
-
-// this is a variable so that it can be overridden in tests
-var getRandomFloat64 = func(value_min, value_max int) float64 {
-	// #nosec G404 - we don't need a cryptographically strong random number generator here
-	return rand.Float64()*(float64(value_max)-float64(value_min)) + float64(value_min)
-}
-
-func (g *hostMetricsGenerator) generateMetrics() pmetric.Metrics {
-	// Update the data points with new values
-	for _, updater := range g.dataPointUpdaters {
-		updater()
-	}
-	return g.otlpGenerator.generateMetrics()
 }
