+++ receiver/telemetrygeneratorreceiver/testdata/expected_metrics/metrics.yaml
